AI/cnn/requirements.txt
fastapi
uvicorn
pillow
torch==2.1.0+cpu
torchvision==0.16.0+cpu